+++ pylsp/plugins/_rope_task_handle.py
@@ -1,32 +1,16 @@
 from __future__ import annotations
 
 import logging
-import time
-from functools import wraps
 
 from typing import Callable, ContextManager, List, Optional, Sequence
 
 from rope.base.taskhandle import BaseJobSet, BaseTaskHandle
 
 from pylsp.workspace import Workspace
+from pylsp._utils import throttle
 
 log = logging.getLogger(__name__)
 Report = Callable[[str, int], None]
-
-
-def throttle(seconds=1):
-    def decorator(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):  # pylint: disable=inconsistent-return-statements
-            if not hasattr(wrapper, "last_call"):
-                wrapper.last_call = 0
-            if time.time() - wrapper.last_call >= seconds:
-                wrapper.last_call = time.time()
-                return func(*args, **kwargs)
-
-        return wrapper
-
-    return decorator
 
 
 class PylspJobSet(BaseJobSet):
